requirements.txt
click==8.1.7
gitdb==4.0.11
GitPython==3.1.43
Jinja2==3.1.3
MarkupSafe==2.1.5
PyYAML==6.0.1
smmap==5.0.1
requests==2.31.0
boto3==1.34.101
packaging==24.0
python-dotenv==1.0.1
inquirer==3.4.0
psutil==6.1.0